+++ pcs/cli/common/lib_wrapper.py
@@ -373,6 +373,9 @@ def load_module(env, middleware_factory, name):
             {
                 "create": stonith.create,
                 "create_in_group": stonith.create_in_group,
+                "history_get_text": stonith.history_get_text,
+                "history_cleanup": stonith.history_cleanup,
+                "history_update": stonith.history_update,
             }
         )
 
@@ -408,7 +411,7 @@ def load_module(env, middleware_factory, name):
 
     raise Exception("No library part '{0}'".format(name))
 
-class Library(object):
+class Library():
     def __init__(self, env, middleware_factory):
         self.env = env
         self.middleware_factory = middleware_factory