+++ pcs/cli/common/test/test_console_report.py
@@ -3069,3 +3069,19 @@ class ResourceBundleUnsupportedContainerType(NameBuildTest):
                 "bundle id", ["b", "a", "c"]
             ),
         )
+
+class FenceHistoryCommandError(NameBuildTest):
+    code = codes.FENCE_HISTORY_COMMAND_ERROR
+    def test_success(self):
+        self.assert_message_from_report(
+            "Unable to show fence history: reason",
+            reports.fence_history_command_error("reason", "show")
+        )
+
+class FenceHistoryNotSupported(NameBuildTest):
+    code = codes.FENCE_HISTORY_NOT_SUPPORTED
+    def test_success(self):
+        self.assert_message_from_report(
+            "Fence history is not supported, please upgrade pacemaker",
+            reports.fence_history_not_supported()
+        )